@@ -1,25 +1,24 @@
 <template>
   <div id="app">
-    
-    <link rel="stylesheet" href="@/assets/projects/projects.css" type="text/css">
-
     <Header />
     <div class="main">
-      <transition name="fade" mode="out-in">
-        <router-view/>
-      </transition>
+      <router-view v-slot="{ Component }">
+        <transition>
+          <component :is="Component" />
+        </transition>
+      </router-view>
     </div>
     <Footer />
   </div>
 </template>
 
 <script lang="ts">
-import Vue from 'vue';
+import { defineComponent } from 'vue';
 import Header from './components/Header.vue';
 import Footer from './components/Footer.vue';
 import Helpers from './helpers';
 
-export default Vue.extend({
+export default defineComponent({
   name: 'App',
   components: {
     Header, Footer
