@@ -1,45 +1,45 @@
 <template>
-    <div>
-      <div class="projects-list">
-        <template v-for="project in projects">
-          <div
-            :key="project.id"
-              @click="showDetails(project)"
-              class="project-item"
-              :class="{ 'wide': project.isWide, 'high': project.isHigh }">
-            <div class="project-item-image" :style="{ 'background-image': 'url(' + project.iconUrl + ')' }">
-            </div>
-            <div class="title-bar" :style="{ 'background-color': project.accentColor + 'DD' }">
-                <div class="title-text">
-                  {{ project.name }}
-                </div>
-              </div>
+  <div>
+    <div class="projects-list">
+      <template v-for="project in projects" :key="project.id"
+     :project="project">
+        <div
+            @click="showDetails(project)"
+            class="project-item"
+            :class="{ 'wide': project.isWide, 'high': project.isHigh }">
+          <div class="project-item-image" :style="{ 'background-image': 'url(' + project.iconUrl + ')' }">
           </div>
-        </template>
-      </div>
-
-      <ProjectDetailsOverlay
-        v-on:close="showPopup = false"
-        :visible="showPopup"
-        :title="popupTitle"
-        :htmlContent="popupContent"
-        :color="popupColor"
-      />
+          <div class="title-bar" :style="{ 'background-color': project.accentColor + 'DD' }">
+              <div class="title-text">
+                {{ project.name }}
+              </div>
+            </div>
+        </div>
+      </template>
     </div>
+
+    <ProjectDetailsOverlay
+      v-on:close="showPopup = false"
+      :visible="showPopup"
+      :title="popupTitle"
+      :htmlContent="popupContent"
+      :color="popupColor"
+    />
+  </div>
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 import ProjectDetailsOverlay from "@/components/ProjectDetailsOverlay.vue";
-import ProjectData from "@/data/ProjectData.ts";
+import ProjectData from "@/data/ProjectData";
 
-export default Vue.extend({
+export default defineComponent({
   name: "ProjectsList",
   components: {
     ProjectDetailsOverlay,
   },
   props: {
-    projects: Array
+    projects: Array<ProjectData>
   },
   data: function () {
     return {
@@ -89,7 +89,7 @@ export default Vue.extend({
 }
 
 .project-item:hover {
-filter: brightness(120%);
+  filter: brightness(120%);
 }
 
 .title-bar {
