@@ -1,10 +1,11 @@
-import Vue from 'vue'
+import { createApp } from 'vue';
 import App from './App.vue'
-import router from './router'
+import { createRouter } from './router'
 
-Vue.config.productionTip = false
+const router = createRouter();
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+const app = createApp(App);
+
+app.use(router);
+
+app.mount('#app');
