@@ -10,11 +10,11 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import {defineComponent} from "vue";
 import ProjectsList from "@/components/ProjectsList.vue";
-import otherProjectsData from "@/data/OtherProjectsData.ts";
+import otherProjectsData from "@/data/OtherProjectsData";
 
-export default Vue.extend({
+export default defineComponent({
   name: "OtherProjects",
   components: {
     ProjectsList,
