@@ -1,48 +1,47 @@
-import Vue from 'vue'
-import VueRouter, { RouteConfig } from 'vue-router'
-import Home from '../views/Home.vue'
+import { createRouter as _createRouter, createWebHistory } from 'vue-router'
 
-Vue.use(VueRouter)
+// route-level code splitting
+const AboutView = () => import('../views/About.vue');
+const ResumeView = () => import('../views/Resume.vue')
+const GameView = () => import('../views/GameProjects.vue')
+const OtherView = () => import('../views/OtherProjects.vue')
+const ContactView = () => import('../views/Contact.vue')
+const E404View = () => import('../views/404.vue')
 
-const routes: Array<RouteConfig> = [
-  {
-    path: '/',
-    name: 'Root',
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  },
-  {
-    path: '/resume',
-    name: 'Resume',
-    component: () => import(/* webpackChunkName: "about" */ '../views/Resume.vue')
-  },
-  {
-    path: '/game-projects',
-    name: 'Game Projects',
-    component: () => import(/* webpackChunkName: "about" */ '../views/GameProjects.vue')
-  },
-  {
-    path: '/other-projects',
-    name: 'Other Projects',
-    component: () => import(/* webpackChunkName: "about" */ '../views/OtherProjects.vue')
-  },
-  {
-    path: '/contact',
-    name: 'Contact',
-    component: () => import(/* webpackChunkName: "about" */ '../views/Contact.vue')
-  },
-  {
-    path: '/404',
-    name: 'NotFound',
-    component: () => import(/* webpackChunkName: "about" */ '../views/404.vue')
-  },
-  {
-    path: '*',
-    redirect: '/404'
-  }
-]
-
-const router = new VueRouter({
-  routes
-})
-
-export default router
+export function createRouter(){
+  return _createRouter({
+    history: createWebHistory(),
+    routes: [
+      {
+        path: '/',
+        name: 'Root',
+        component: AboutView
+      },
+      {
+        path: '/resume',
+        name: 'Resume',
+        component: ResumeView
+      },
+      {
+        path: '/game-projects',
+        name: 'Game Projects',
+        component: GameView
+      },
+      {
+        path: '/other-projects',
+        name: 'Other Projects',
+        component: OtherView
+      },
+      {
+        path: '/contact',
+        name: 'Contact',
+        component: ContactView
+      },
+      {
+        path: '/:pathMatch(.*)',
+        name: 'NotFound',
+        component: E404View
+      }
+    ]
+  })
+}
